Namespaces middleware to avoid conflicts with existing middleware in app named the same.
